user 구조 변경
